+++ src/common/classes/app-pages-base.class.js
@@ -1,0 +1,30 @@
+/*
+ * pwix:app-pages/src/common/classes/app-pages-base.class.js
+ *
+ * A common base class.
+ * This is needed as '@vestergaard-company' interface mechanism implies to derivate a class from another class.
+ */
+
+import _ from 'lodash';
+import { strict as assert } from 'node:assert';
+
+export class AppPagesBase {
+
+    // static data
+
+    // static methods
+
+    // private data
+
+    // private methods
+
+    // public data
+
+    /**
+     * @locus Anywhere
+     * @constructor
+     */
+    constructor(){
+        return this;
+    }
+}