+++ src/common/js/index.js
@@ -8,6 +8,8 @@ import { MenuSet } from '../classes/menu-set.class.js';
 import { MenuItem } from '../classes/menu-item.class.js';
 import { RunContext } from '../classes/run-context.class.js';
 
+import { IMenuItem } from '../interfaces/imenu-item.iface.js';
+
 import './global.js';
 import './constants.js';
 import './configure.js';
@@ -18,3 +20,5 @@ AppPages.DisplayUnit = DisplayUnit;
 AppPages.MenuSet = MenuSet;
 AppPages.MenuItem = MenuItem;
 AppPages.RunContext = RunContext;
+
+AppPages.IMenuItem = IMenuItem;